requirements.txt
pyTelegramBotAPI>=4.19.0
pyYAML>=3.12
